requirements.txt
pandas
pyyaml
pydantic